refactor: move useBlogErrorToast to the hooks directory
