+++ frontend/src/routes/blog/index.tsx
@@ -9,6 +9,7 @@ import { BlogFilter } from "~/features/Blog/BlogFilter";
 import { BlogHeader } from "~/features/Blog/BlogHeader";
 import BlogPostList from "~/features/Blog/BlogPostList";
 import BlogSidebar from "~/features/Blog/BlogSidebar";
+import { useBlogErrorToast } from "~/hooks/useBlogErrorToast";
 import { PostState } from "~/models/post";
 import {
   availableActionsQuery,
@@ -20,7 +21,6 @@ import {
   usePostsList,
 } from "~/services/queries";
 import { useStoreUpdaters } from "~/store";
-import { useBlogErrorToast } from "~/utils/BlogEvent";
 
 export const blogLoader =
   (queryClient: QueryClient) =>
